Move PostUpdateSets to common; add FootIk + Nametag variants
Track foot-IK and nametag PostUpdate ordering points as variants of the
shared PostUpdateSets enum (now in common::sets, alongside SceneSets)
rather than ad-hoc per-feature SystemSet structs, so the full PostUpdate
ordering chain is visible in one place. Producers slot into the variant
via `.in_set(...)`; the chain in TransformAndParentPlugin handles the
cross-stage ordering.
